Get rid of duplicated code

+++ server/src/main/java/org/sonarsource/ldap/server/ApacheDS.java
@@ -1,5 +1,5 @@
 /*
- * SonarQube LDAP Plugin :: Integration Tests
+ * SonarQube LDAP Test Server
  * Copyright (C) 2009-2016 SonarSource SA
  * mailto:contact AT sonarsource DOT com
  *
@@ -17,20 +17,36 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonarsource.ldap.it;
+package org.sonarsource.ldap.server;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import com.google.common.io.Closeables;
+import java.io.File;
+import java.io.InputStream;
+import java.util.Collections;
+import java.util.Hashtable;
+import java.util.Map;
+import javax.naming.Context;
+import javax.naming.directory.Attribute;
+import javax.naming.directory.Attributes;
+import javax.naming.directory.BasicAttribute;
+import javax.naming.directory.DirContext;
+import javax.naming.directory.InitialDirContext;
+import javax.naming.directory.ModificationItem;
+import javax.naming.ldap.InitialLdapContext;
 import org.apache.commons.io.FileUtils;
+import org.apache.directory.server.constants.ServerDNConstants;
 import org.apache.directory.server.core.CoreSession;
 import org.apache.directory.server.core.DefaultDirectoryService;
 import org.apache.directory.server.core.DirectoryService;
 import org.apache.directory.server.core.entry.DefaultServerEntry;
 import org.apache.directory.server.core.entry.ServerEntry;
+import org.apache.directory.server.core.jndi.CoreContextFactory;
 import org.apache.directory.server.core.partition.impl.btree.jdbm.JdbmIndex;
 import org.apache.directory.server.core.partition.impl.btree.jdbm.JdbmPartition;
+import org.apache.directory.server.kerberos.kdc.KdcServer;
 import org.apache.directory.server.ldap.LdapServer;
 import org.apache.directory.server.ldap.handlers.bind.MechanismHandler;
 import org.apache.directory.server.ldap.handlers.bind.cramMD5.CramMd5MechanismHandler;
@@ -38,6 +54,7 @@ import org.apache.directory.server.ldap.handlers.bind.digestMD5.DigestMd5Mechani
 import org.apache.directory.server.ldap.handlers.bind.gssapi.GssapiMechanismHandler;
 import org.apache.directory.server.ldap.handlers.bind.plain.PlainMechanismHandler;
 import org.apache.directory.server.protocol.shared.transport.TcpTransport;
+import org.apache.directory.server.protocol.shared.transport.UdpTransport;
 import org.apache.directory.server.xdbm.Index;
 import org.apache.directory.shared.ldap.constants.SupportedSaslMechanisms;
 import org.apache.directory.shared.ldap.ldif.ChangeType;
@@ -47,24 +64,26 @@ import org.apache.mina.util.AvailablePortFinder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.File;
-import java.io.InputStream;
-import java.util.Map;
-
-/**
- * This class originally was developed in sonar-ldap-plugin and then was duplicated here with some modifications.
- */
-public class ApacheDS {
+public final class ApacheDS {
 
   private static final Logger LOG = LoggerFactory.getLogger(ApacheDS.class);
 
-  public static ApacheDS start(String baseDn, String workDir) throws Exception {
-    return new ApacheDS(baseDn)
-        .startDirectoryService(workDir)
-        .startLdapServer();
+  private final String realm;
+  private final String baseDn;
+
+  public static ApacheDS start(String realm, String baseDn, String workDir) throws Exception {
+    return new ApacheDS(realm, baseDn)
+      .startDirectoryService(workDir)
+      .startLdapServer()
+      .activateNis();
+  }
+
+  public static ApacheDS start(String realm, String baseDn) throws Exception {
+    return start(realm, baseDn, "target/ldap-work/" + realm);
   }
 
   public void stop() throws Exception {
+    // kdcServer.stop();
     ldapServer.stop();
     directoryService.shutdown();
   }
@@ -90,6 +109,8 @@ public class ApacheDS {
           rootDSE.modify(ldifEntry.getDn(), ldifEntry.getModificationItems());
         } else if (ChangeType.Delete == ldifEntry.getChangeType()) {
           rootDSE.delete(ldifEntry.getDn());
+        } else {
+          throw new IllegalStateException();
         }
       }
     } finally {
@@ -107,14 +128,16 @@ public class ApacheDS {
     ldapServer.setAllowAnonymousAccess(true);
   }
 
-  private final String baseDn;
   private final DirectoryService directoryService;
   private final LdapServer ldapServer;
+  private final KdcServer kdcServer;
 
-  private ApacheDS(String baseDn) {
+  private ApacheDS(String realm, String baseDn) {
+    this.realm = realm;
     this.baseDn = baseDn;
     directoryService = new DefaultDirectoryService();
     ldapServer = new LdapServer();
+    kdcServer = new KdcServer();
   }
 
   private ApacheDS startDirectoryService(String workDirStr) throws Exception {
@@ -123,7 +146,9 @@ public class ApacheDS {
     directoryService.setShutdownHookEnabled(false);
 
     File workDir = new File(workDirStr);
-    FileUtils.deleteDirectory(workDir);
+    if (workDir.exists()) {
+      FileUtils.deleteDirectory(workDir);
+    }
     directoryService.setWorkingDirectory(workDir);
 
     JdbmPartition partition = new JdbmPartition();
@@ -158,12 +183,74 @@ public class ApacheDS {
     ldapServer.setSaslMechanismHandlers(mechanismHandlerMap);
 
     ldapServer.setSaslHost("localhost");
-    // ldapServer.setSaslRealms(Collections.singletonList("example.org"));
+    ldapServer.setSaslRealms(Collections.singletonList(realm));
     // TODO ldapServer.setSaslPrincipal();
     // The base DN containing users that can be SASL authenticated.
     ldapServer.setSearchBaseDn(baseDn);
 
     ldapServer.start();
+
+    return this;
+  }
+
+  @SuppressWarnings("unused")
+  private ApacheDS startKerberos() throws Exception {
+    Preconditions.checkState(ldapServer.isStarted());
+
+    kdcServer.setDirectoryService(directoryService);
+    // FIXME hard-coded ports
+    kdcServer.setTransports(new TcpTransport(6088), new UdpTransport(6088));
+    kdcServer.setEnabled(true);
+    kdcServer.setPrimaryRealm(realm);
+    kdcServer.setSearchBaseDn(baseDn);
+    kdcServer.setKdcPrincipal("krbtgt/" + realm + "@" + baseDn);
+    kdcServer.start();
+
+    // -------------------------------------------------------------------
+    // Enable the krb5kdc schema
+    // -------------------------------------------------------------------
+
+    Hashtable<String, Object> env = new Hashtable<>();
+    env.put(DirectoryService.JNDI_KEY, directoryService);
+    env.put(Context.INITIAL_CONTEXT_FACTORY, CoreContextFactory.class.getName());
+    env.put(Context.PROVIDER_URL, ServerDNConstants.OU_SCHEMA_DN);
+    InitialLdapContext schemaRoot = new InitialLdapContext(env, null);
+
+    // check if krb5kdc is disabled
+    Attributes krb5kdcAttrs = schemaRoot.getAttributes("cn=Krb5kdc");
+    boolean isKrb5KdcDisabled = false;
+    if (krb5kdcAttrs.get("m-disabled") != null) {
+      isKrb5KdcDisabled = ((String) krb5kdcAttrs.get("m-disabled").get()).equalsIgnoreCase("TRUE");
+    }
+
+    // if krb5kdc is disabled then enable it
+    if (isKrb5KdcDisabled) {
+      Attribute disabled = new BasicAttribute("m-disabled");
+      ModificationItem[] mods = new ModificationItem[] {new ModificationItem(DirContext.REMOVE_ATTRIBUTE, disabled)};
+      schemaRoot.modifyAttributes("cn=Krb5kdc", mods);
+    }
+    return this;
+  }
+
+  /**
+   * This seems to be required for objectClass posixGroup.
+   */
+  private ApacheDS activateNis() throws Exception {
+    Preconditions.checkState(ldapServer.isStarted());
+
+    Attribute disabled = new BasicAttribute("m-disabled", "TRUE");
+    Attribute disabled2 = new BasicAttribute("m-disabled", "FALSE");
+    ModificationItem[] mods = new ModificationItem[] {
+      new ModificationItem(DirContext.REMOVE_ATTRIBUTE, disabled),
+      new ModificationItem(DirContext.ADD_ATTRIBUTE, disabled2)
+    };
+
+    Hashtable env = new Hashtable();
+    env.put(Context.INITIAL_CONTEXT_FACTORY, "com.sun.jndi.ldap.LdapCtxFactory");
+    env.put(Context.PROVIDER_URL, getUrl());
+
+    DirContext ctx = new InitialDirContext(env);
+    ctx.modifyAttributes("cn=nis,ou=schema", mods);
 
     return this;
   }
